src/requirements.txt
bcrypt==3.1.7
cffi==1.13.1
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Session==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysql-connector==2.2.9
numpy==1.17.2
pandas==0.25.1
pycparser==2.19
pymongo==3.9.0
python-dateutil==2.8.0
pytz==2019.3
six==1.12.0
Werkzeug==0.16.0
